Added some functionality
Users can now define what their non-routable VLAN is via arguments. Cleaned up some public-facing debug statements. Invalid certificates are now accepted, but generate a log for the user to review. User can now see what interfaces were changed. Ensured that interfaces that are being modified are configured to be switchports before applying switchport configuration. Modified the payload version, as 1.2 wasn't working on the chassis I was testing on (not sure if that's dependent on the version of NX-OS or not, but it probably is.) Added command-line switches that allow for the choice between interfaces that are not connected or missing SFPs (by default, script adds configuration to both, whereas previously it only affected interfaces with missing SFPs.)